refactor: separate types in external file
