+++ client/src/pages/GeneratorPage/GeneratorPage.tsx
@@ -3,16 +3,8 @@ import { SubmitHandler, useForm } from 'react-hook-form';
 import NavOutlet from '../../components/NavOutlet';
 import GenerateForm from '../../components/GenerateForm/GenerateForm';
 import GeneratedActivity from '../../components/GeneratedActivity/GeneratedActivity';
+import { IFormInput } from '../../types/activity';
 import './GeneratorPage.css';
-
-export interface IFormInput {
-  Topic: {};
-  KidsNumber: {};
-  AgeRange: {};
-  Difficulty: {};
-  Place: {};
-  Duration: {};
-}
 
 const GeneratorPage = () => {
   const { control, handleSubmit } = useForm<IFormInput>({});
@@ -21,9 +13,8 @@ const GeneratorPage = () => {
 
   const onSubmit: SubmitHandler<IFormInput> = (data) => {
     setLoading(true);
-    const apiEndpoint = 'http://localhost:3000/generator';
-    console.log('data', data);
-    fetch(apiEndpoint, {
+
+    fetch('http://localhost:3000/generator', {
       method: 'POST',
       headers: {
         'Content-Type': 'application/json',
@@ -34,12 +25,16 @@ const GeneratorPage = () => {
         return response.json();
       })
       .then((dataReceived) => {
-        console.log('dataReceived', dataReceived);
         // If working with openAI, use:
         // setActivity(dataReceived);
+
         // If working with MOCK data, use:
-        setActivity(dataReceived.openAIResponse.content);
-        setLoading(false);
+        if (dataReceived.matchingActivity!) {
+          setActivity(dataReceived.matchingActivity);
+          setLoading(false);
+        } else {
+          console.error('No match have been found in mock data');
+        }
       })
       .catch((error) => {
         console.error('Error:', error);
